cgroup2 support
Refactored how the Dockside container ID is identified. The method for
systems running cgroup v1 is not applicable on systems running cgroup v2,
unless the --cgroupns=host option to 'docker run' is used when launching.

For launching Dockside on systems with cgroup v2 without --cgroupns=host,
Dockside now uses alternative methods for identifying the Dockside
container including:
- docker run --cidfile option passing container id to bind-mounted fs
- obtaining the container's hostname from $HOSTNAME

N.B. The Dockside container ID is required to identify the IDE volume
when bind-mounting /var/run/docker.dock, but is not required when
running an inner dockerd, e.g. with the Sysbox runtime.